add a defaultEmptyComponent method  for link and image directives
