+++ packages/sitecore-jss-angular/src/components/date.directive.ts
@@ -46,7 +46,7 @@ export class DateDirective extends BaseFieldDirective implements OnChanges {
     protected renderer: Renderer2,
     protected elementRef: ElementRef
   ) {
-    super(viewContainer, renderer, elementRef);
+    super(viewContainer);
     this.defaultFieldEditingComponent = DefaultEmptyFieldEditingComponent;
   }
