+++ packages/sitecore-jss-angular/src/components/text.directive.ts
@@ -40,7 +40,7 @@ export class TextDirective extends BaseFieldDirective implements OnChanges {
     protected renderer: Renderer2,
     protected elementRef: ElementRef
   ) {
-    super(viewContainer, renderer, elementRef);
+    super(viewContainer);
     this.defaultFieldEditingComponent = DefaultEmptyFieldEditingComponent;
   }
